Reorganized things, added health endpoint
