+++ models/email_models.go
@@ -1,4 +1,4 @@
-package main
+package models
 
 // EMAIL represents an email address to be added to our mailing list.
 type Email struct {
